feat(featured-listings): wire real store data into FeaturedListings
- Make home page async to fetch RECOMMENDED stores server-side via getHighlightedStores()
- Replace hardcoded mock cards with live Firestore data; section hides when no approved stores exist
- Extract storeToCard transformer into utils/store.ts and reuse it in ResultsArea
- Remove decorative pill filter row from FeaturedListings
